+++ packages/python-ta/CHANGELOG.md
@@ -17,6 +17,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### 💫 New checkers
 
 - `invalid-precondition-syntax`: Added new checker that checks if a function contains invalid syntax within its precondition statements.
+- `invalid-postcondition-syntax`: Added new checker that checks if a function contains invalid syntax within its postcondition statements.
 
 ### 🐛 Bug fixes